+++ android/src/main/java/br/com/rsmarques/flutter_branch_sdk/FlutterBranchSdkPlugin.java
@@ -237,6 +237,7 @@ public class FlutterBranchSdkPlugin implements FlutterPlugin, MethodCallHandler,
       case "getShortUrl":
         getShortUrl(call, result);
         break;
+      case "shareWithLPLinkMetadata":
       case "showShareSheet":
         showShareSheet(call, result);
         break;
@@ -296,6 +297,9 @@ public class FlutterBranchSdkPlugin implements FlutterPlugin, MethodCallHandler,
         break;
       case "getQRCode":
         getQRCode(call, result);
+        break;
+      case "handleDeepLink":
+        handleDeepLink(call);
         break;
       default:
         result.notImplemented();
@@ -750,6 +754,22 @@ public class FlutterBranchSdkPlugin implements FlutterPlugin, MethodCallHandler,
         result.success(response);
       }
     }
+  private void handleDeepLink(final MethodCall call) {
+
+    LogUtils.debug(DEBUG_NAME, "handleDeepLink call");
+    if (!(call.arguments instanceof Map)) {
+      throw new IllegalArgumentException("Map argument expected");
+    }
+    HashMap<String, Object> argsMap = (HashMap<String, Object>) call.arguments;
+
+    final String url = call.argument("url");
+
+    Intent intent = new Intent(context, activity.getClass());
+    intent.putExtra("branch",url);
+    intent.putExtra("branch_force_new_session",true);
+    activity.startActivity(intent);
+  }
+
 }