+++ ios/Classes/FlutterBranchIoSdkHelper.swift
@@ -263,9 +263,9 @@ extension Bundle {
     }
     public var icon: UIImage? {
         if let icons = infoDictionary?["CFBundleIcons"] as? [String: Any],
-            let primaryIcon = icons["CFBundlePrimaryIcon"] as? [String: Any],
-            let iconFiles = primaryIcon["CFBundleIconFiles"] as? [String],
-            let lastIcon = iconFiles.last {
+           let primaryIcon = icons["CFBundlePrimaryIcon"] as? [String: Any],
+           let iconFiles = primaryIcon["CFBundleIconFiles"] as? [String],
+           let lastIcon = iconFiles.last {
             return UIImage(named: lastIcon)
         }
         return nil
@@ -315,5 +315,5 @@ extension UIImage {
             }
         }
     }
-
+    
 }